code/src/back_end/requirements.txt
openai>=1.0.0
python-dotenv>=0.20.0 